backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
numpy==1.24.3
pandas==2.0.3
opencv-python==4.8.1.78
scikit-image==0.21.0
tensorflow==2.14.0
python-multipart==0.0.6
pillow==10.0.1
scikit-image